requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
setuptools==80.9.0
mysqlclient==2.2.0 
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
mypy==1.7.1
types-python-dateutil==2.8.19.14 